Split session file IPC handler
